Split SIF and RPC logic
