+++ lib/get-contacts.js
@@ -1,0 +1,6 @@
+module.exports = (contacts, role) => {
+  if (contacts && Array.isArray(contacts)) {
+    const { SearchName } = contacts.find(contact => contact.Role === role) || {}
+    return SearchName
+  }
+}
